Rename recursiveSortingTrait.php to RecursiveSortingTrait.php Change use statement from `use recursiveSortingTrait` to `use RecursiveSortingTrait`
